separated portal into smaller components and fixed some bugs
